use spline interpolation for pid control
The previous command was generated by finding the next goal and going straight ahead. Instead, this new method uses a spline interpolation using 3 points (current position, goal, next goal)
and the current angular position to generate a smoother curve.
Mind that this is still WIP and does not work completely (eg. the system has a tendency to bump into walls and stay stuck in there) - this will be fixed in the future commits